LayoutTests/http/tests/security/contentSecurityPolicy/1.1/stylenonce-allowed-expected.txt

CONSOLE MESSAGE: Refused to apply inline style because it violates the following Content Security Policy directive: "style-src 'nonce-noncynonce' 'nonce-noncy+/=nonce'". Either the 'unsafe-inline' keyword, a hash ('sha256-SKwGvORdKBYTYiM4lxIkanDyKH8J0qJ5Ix8LGkKsbhw='), or a nonce ('nonce-...') is required to enable inline execution.

CONSOLE MESSAGE: line 11: Refused to apply inline style because it violates the following Content Security Policy directive: "style-src 'nonce-noncynonce' 'nonce-noncy+/=nonce'". Either the 'unsafe-inline' keyword, a hash ('sha256-kv95ImKKneBhnSXrPlx5XNiVbPjFnuiudpQxG+M00io='), or a nonce ('nonce-...') is required to enable inline execution.

Style correctly whitelisted via a 'nonce-*' expression in 'style-src' should be applied to the page.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.getComputedStyle(el).color is "rgb(0, 128, 0)"
PASS window.getComputedStyle(el).color is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
This text should be green.

This text should also be green.
